Disregard any ESLint stderr to calculate exit code
Since the goal of eslines is to downgrade errors on non-modified lines,
we can't count on eslint's exit code. Previously, we assumed that if
eslint outputs something on its stderr, this means it encountered some
other error. It turned out to be a bad assumption, as the following
example illustrates.

Currently, the eslint-plugin-react outputs some warning in the stderr
channel https://github.com/interfirm/eslint-config-interfirm/issues/57
ESLint will output that errors in the stderr but if it doesn't find any
lint issues it will report zero as its exit code. In our case,
although eslines will exit 0 as well, we'd return 1 because
we heard ESLint reported something in the stderr channel.